dot product in df
How do I get the dot product of a column in a dataframe, instead of creating another pickle with just the title vectors?
 
EG, now we have a df with 2 columns. We have this code working for 1-D vectors. How do I apply this between just the “vector columns” in these 2 dataframes (other titles, vectorized titles?)